Rename `SlotPool` -> `SlotPoolImpl` and `SlotPoolGateway` -> `SlotPool`
